Ajustes

+++ Atividade_3/src/fs.h
@@ -13,7 +13,7 @@ public:
     static const unsigned short int POINTERS_PER_INODE = 5;
     static const unsigned short int POINTERS_PER_BLOCK = 1024;
     static const unsigned short int POINTERS_TO_BLOCKS_PER_INODE = POINTERS_PER_INODE + POINTERS_PER_BLOCK;
-    static const unsigned short int BYTES_PER_INODE = POINTERS_TO_BLOCKS_PER_INODE * Disk::DISK_BLOCK_SIZE;
+    static const int BYTES_PER_INODE = POINTERS_TO_BLOCKS_PER_INODE * Disk::DISK_BLOCK_SIZE;
 
     class fs_superblock {
         public:
